fix: revert agentCliSessionStores rename, document dual ownership instead
Automated review on the PR caught that the earlier rename to
agentShutdownHandlers.ts was based on a false premise: the module isn't
just shutdown wiring, it owns the two session-keyed WeakMap stores
(codexThreadsFor, claudeAgentSessionsFor) that the rest of the codebase
already calls the agent-CLI "session store" (AgentCliSessionStoreAccessor
in agentCliShared.ts, used for live registration/lookup in codex.ts and
claudeAgent.ts, not just teardown). Renaming it away from "Stores" made it
undiscoverable by the vocabulary already in use elsewhere.

Revert the filename to agentCliSessionStores.ts and instead add a module
docstring recording the file's actual dual responsibility (session-store
WeakMaps + shutdown-handler registration), so the real overlap is
documented rather than hidden behind a name that described only half of
it. The controllers/session/createSessionStores.ts rename is unaffected —
review confirmed it correctly resolves the case-only collision with
src/agent/storage/SessionStores.ts.
